requirements.txt
openai==2.30.0
qdrant-client==1.17.1
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-chroma==0.2.4
langchain-community==0.3.24
chromadb==1.0.9
pymupdf==1.25.5
fastapi==0.115.9
uvicorn[standard]==0.34.2
streamlit==1.45.0
python-dotenv==1.0.1
tqdm==4.67.1
pydantic==2.10.6
redis==7.4.0 
SQLAlchemy==2.0.49
psycopg2-binary==2.9.11